backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.9.2
pydantic-settings>=2.5.2
websockets>=13.1
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
slowapi>=0.1.9
networkx>=3.3
httpx>=0.27.2
python-multipart>=0.0.12
# Google ADK (optional — agents fall back to deterministic reasoning if missing)
google-adk>=0.2.0
google-generativeai>=0.8.3
# tests
pytest>=8.3.3
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0
locust>=2.31.0
